Add uniqueness constraint on replications for (study_id, replicating_study_id)
I had to add another study in the tests since let! creates the record
and doesn't roll it back between test runs.  This was causing uniqueness
to be violated.

I also had to make Study.add_replication use .create! so that an error
would be raised if the validations fail.